Split api_utils to several modules
