+++ microchat/api_utils/request.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class APIRequest:
+    pass
